// NhapKhoApp/src/generated/models/Htn1423_nhapkhosModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Htn1423_nhapkhosstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Htn1423_nhapkhosstatecode = keyof typeof Htn1423_nhapkhosstatecode;
export const Htn1423_nhapkhosstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Htn1423_nhapkhosstatuscode = keyof typeof Htn1423_nhapkhosstatuscode;

export interface Htn1423_nhapkhosBase {
  htn1423_name: string;
  htn1423_nhapkhoid: string;
  "htn1423_onmuachitiet@odata.bind": string;
  htn1423_soluong: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Htn1423_nhapkhosstatecode;
  statuscode?: Htn1423_nhapkhosstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Htn1423_nhapkhos extends Htn1423_nhapkhosBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  htn1423_onmuachitietname?: string;
  htn1423_tensanpham?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  htn1423_onmuachitiet?: object;
  _htn1423_onmuachitiet_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}